requirements.txt
pandas==1.4.2
Pillow==9.2.0
plotly==5.6.0
streamlit==1.12.0
